Category View + simple "add new task"
Hey guys! I did the category view, it has dummy numbers for now since we need other features to collect data for those statistics. So, I will do the math for the donut chart and stuff when we get more work done. Also, I made a temporary, and simple "add a new task" feature at the bottom that might help whoever's working on that part. If not, I can finish it tomorrow! :)