Implemented 'changing the order of todo items'
Implemented a combination of drag-and-drop and multiselect features using the 'react-native-multiselect-sortable-flatlist' library.

Changes the return elements in each sorting function in SortAs.js, as it was creating unnecessary documents within the database that contained only 'completed: true/false' because of the lack of an id to locate the item. It should be fine now.